v1.2.0 - Updated Projects
