+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "john-dufresne-portfolio",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
